+++ api/programs/test-meta/src/wasm.rs
@@ -6,7 +6,7 @@ static mut STATE: Vec<Wallet> = Vec::new();
 
 #[no_mangle]
 unsafe extern "C" fn init() {
-    let message_in: BTreeSet<u8> = msg::load().expect("Unable to decode message");
+    let _: BTreeSet<u8> = msg::load().expect("Unable to decode message");
 
     let mut res: BTreeMap<String, u8> = BTreeMap::new();
 
@@ -42,12 +42,15 @@ unsafe extern "C" fn init() {
 
     msg::reply(res, 0).unwrap();
 }
+
 #[no_mangle]
 unsafe extern "C" fn handle() {
     let action: Action = msg::load().unwrap();
 
     match action {
-        Action::One(option_string) => msg::reply(EmptyStruct { empty: () }, 1000).unwrap(),
+        Action::One(_) => {
+            msg::send_with_gas(msg::source(), EmptyStruct { empty: () }, 10000000, 1000).unwrap()
+        }
         _ => msg::reply(EmptyStruct { empty: () }, 0).unwrap(),
     };
 }
@@ -55,10 +58,4 @@ unsafe extern "C" fn handle() {
 #[no_mangle]
 extern "C" fn state() {
     msg::reply(unsafe { STATE.clone() }, 0).expect("Error in state");
-}
-
-#[no_mangle]
-extern "C" fn metahash() {
-    let metahash: [u8; 32] = include!("../.metahash");
-    msg::reply(metahash, 0).expect("Failed to share metahash");
 }